:recycle: chore: refactor Header component
