--- mobile/src/components/Header.tsx
@@ -1,62 +1,0 @@
-import React from 'react';
-import { View, StyleSheet, Text } from 'react-native';
-import { BorderlessButton } from 'react-native-gesture-handler';
-import { Feather } from '@expo/vector-icons';
-import { useNavigation } from '@react-navigation/native';
-
-interface HeaderProps {
-  title: string;
-  showCancel?: boolean;
-}
-
-export default function Header({ title, showCancel = true }: HeaderProps) {
-  const navigation = useNavigation();
-  
-  function handleGoBackToAppHomepage() {
-    navigation.navigate('OrphanagesMap');
-  }
-
-  return (
-    <View style={styles.container}>
-      <BorderlessButton onPress={navigation.goBack}>
-        <Feather name="arrow-left" size={24} color="#15b6d6" />
-      </BorderlessButton>
-
-      <Text style={styles.title}>{title}</Text>
-
-      { showCancel 
-        ? (
-          <BorderlessButton onPress={handleGoBackToAppHomepage}>
-            <Feather name="x" size={24} color="#ff669d" />
-          </BorderlessButton>
-        )
-        : (
-          <View />
-        )
-      }
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    
-    padding: 24,
-    paddingTop: 44,
-
-    backgroundColor: '#f9fafc',
-    
-    borderBottomWidth: 1,
-    borderColor: '#dde3f0',
-  },
-
-  title: {
-    color: '#8fa7b3',
-
-    fontFamily: 'Nunito_600SemiBold',
-    fontSize: 16,
-  }
-});
